Improved second collision snap. Improved ray cast detection

+++ src/gameplay/ball/systems.rs
@@ -1,4 +1,4 @@
-use std::collections::HashMap;
+use std::{collections::HashMap, ops::Add};
 
 use bevy::{
     prelude::{
@@ -17,7 +17,8 @@ use crate::{
     game_audio::utils::pkv_play_shoot_audio,
     gameplay::{
         constants::{
-            BALL_RADIUS, CAST_RAY_MAX_TOI, CAST_RAY_VELOCITY, NEXT_PROJECTILE_SPAWN_BOTTOM,
+            BALL_RADIUS, CAST_RAY_BOUNCE_Y_ADD, CAST_RAY_MAX_TOI, CAST_RAY_TRIES,
+            CAST_RAY_VELOCITY, CAST_RAY_VELOCITY_TOLERANCE, NEXT_PROJECTILE_SPAWN_BOTTOM,
             NEXT_PROJECTILE_SPAWN_SIDE, PROJECTILE_SHOOT_BOTTOM, PROJECTILE_SPAWN_BOTTOM,
             PROJECTILE_SPEED,
         },
@@ -38,7 +39,7 @@ use crate::{
 use super::{
     aim_bundle::AimBundle,
     components::{
-        Aim, AimLine, AimTarget, GridBall, NextProjectileBall, OutBall, ProjectileBall, Species,
+        AimLine, AimTarget, GridBall, NextProjectileBall, OutBall, ProjectileBall, Species,
     },
     projectile_ball_bundle::{NextProjectileBallBundle, ProjectileBallBundle},
     resources::ProjectileBuffer,
@@ -150,13 +151,17 @@ pub fn shoot_projectile(
     audio_assets: Res<AudioAssets>,
     pointer_cooldown: Res<PointerCooldown>,
     pkv: Res<PkvStore>,
-    mut aim_query: Query<&mut Aim, With<Aim>>,
+    mut aim_target_query: Query<&mut AimTarget, With<AimTarget>>,
 ) {
     if pointer_cooldown.started {
         return;
     }
-    let (pointer_position, pointer_aquired) =
+    let (pointer_position, pointer_pressed, pointer_released, pointer_aquired) =
         detect_pointer_position(&window_query, &camera_query, &mouse_button_input, &touches);
+
+    if !(pointer_pressed || pointer_released) {
+        return;
+    }
 
     if let Ok((projectile_transform, mut vel, mut projectile_ball)) =
         projectile_ball_query.get_single_mut()
@@ -170,33 +175,22 @@ pub fn shoot_projectile(
             target_position.y = projectile_shoot_bottom;
         }
 
-        if let Ok(mut aim) = aim_query.get_single_mut() {
-            aim.started = pointer_aquired;
-            if pointer_aquired {
-                if !projectile_ball.is_flying {
+        if let Ok(mut aim) = aim_target_query.get_single_mut() {
+            aim.pointer_pressed = pointer_pressed;
+            aim.pointer_released = pointer_released;
+            if pointer_pressed {
+                if projectile_ball.is_flying {
+                    aim.pointer_pressed = false
+                } else {
                     aim.aim_pos = projectile_transform.translation.truncate();
-                    aim.aim_vel = (target_position - aim.aim_pos).normalize() * CAST_RAY_VELOCITY;
-                    aim.draw_vel = Vec2::ZERO;
-                    // println!(
-                    //     "pointer_aquired target_pos({:?}) projectile_pos({})",
-                    //     target_position, aim.aim_pos
-                    // );
+                    aim.aim_vel = target_position - aim.aim_pos;
                 }
-            } else {
-                aim.aim_pos = Vec2::ZERO;
-                aim.aim_vel = Vec2::ZERO;
             }
         }
 
-        if pointer_aquired && !projectile_ball.is_flying {
-            if !(mouse_button_input.just_released(MouseButton::Left) || touches.any_just_released())
-            {
-                return;
-            }
-
+        if pointer_aquired && !projectile_ball.is_flying && pointer_released {
             let aim_direction = target_position - projectile_transform.translation.truncate();
             vel.linvel = aim_direction.normalize() * PROJECTILE_SPEED;
-            // println!("aim_direction ({}, {})", aim_direction.x, aim_direction.y);
 
             projectile_ball.is_flying = true;
 
@@ -209,90 +203,97 @@ pub fn draw_aim(
     mut commands: Commands,
     mut meshes: ResMut<Assets<Mesh>>,
     gameplay_materials: Res<GameplayMaterials>,
-    mut aim_query: Query<&mut Aim, With<Aim>>,
     wall_query: Query<Entity, With<WallType>>,
     balls_query: Query<Entity, With<GridBall>>,
     rapier_context: Res<RapierContext>,
     aim_line_query: Query<Entity, With<AimLine>>,
-    grid: Res<Grid>,
     lines_query: Query<(&LineType, &Transform), With<LineType>>,
     mut aim_target_query: Query<
-        (&mut Transform, &mut Visibility),
+        (&mut AimTarget, &mut Transform, &mut Visibility),
         (With<AimTarget>, Without<LineType>),
     >,
 ) {
-    if let Ok(mut aim) = aim_query.get_single_mut() {
-        if aim.draw_vel != aim.aim_vel {
-            if let Ok((mut target_transform, mut target_visibility)) =
-                aim_target_query.get_single_mut()
-            {
-                cleanup_aim_line_utils(&mut commands, &aim_line_query);
-                aim.draw_vel = aim.aim_vel;
-                if aim.started {
-                    *target_visibility = Visibility::Visible;
-                    let kinematic_filter = QueryFilter::default().groups(CollisionGroups::new(
-                        Group::GROUP_1 | Group::GROUP_2 | Group::GROUP_5,
-                        Group::GROUP_1 | Group::GROUP_2 | Group::GROUP_5,
-                    ));
+    if let Ok((mut aim_target, mut target_transform, mut target_visibility)) =
+        aim_target_query.get_single_mut()
+    {
+        if aim_target.pointer_released {
+            *target_visibility = Visibility::Hidden;
+            cleanup_aim_line_utils(&mut commands, &aim_line_query);
+            aim_target.draw_vel = Vec2::ZERO;
+        }
 
-                    let shape = Collider::ball(BALL_RADIUS);
-                    let start_y = aim.aim_pos.y;
-                    let mut ray_start = aim.aim_pos;
-                    let mut ray_vel = aim.aim_vel;
-                    let mut count = 0;
+        if aim_target.pointer_pressed && aim_target.draw_vel != aim_target.aim_vel {
+            aim_target.draw_vel = aim_target.aim_vel;
+            // redraw only if pointer position (draw velocity in this case) changed
 
-                    while start_y <= grid.bounds.maxs.y - grid.layout.hex_size.y {
-                        count = count + 1;
-                        if count > 10 {
-                            warn!("Break aim iteration draw, reached max iteration count");
-                            break;
-                        }
-                        if let Some((entity, hit)) = rapier_context.cast_shape(
-                            ray_start,
+            cleanup_aim_line_utils(&mut commands, &aim_line_query);
+            let kinematic_filter = QueryFilter::default().groups(CollisionGroups::new(
+                Group::GROUP_1 | Group::GROUP_2 | Group::GROUP_5,
+                Group::GROUP_1 | Group::GROUP_2 | Group::GROUP_5,
+            ));
+            let shape = Collider::ball(BALL_RADIUS);
+            let mut ray_start = aim_target.aim_pos;
+            let mut ray_vel = aim_target.aim_vel.normalize() * CAST_RAY_VELOCITY;
+            let mut count = 0;
+
+            loop {
+                count = count + 1;
+                if count > CAST_RAY_TRIES {
+                    let message = format!("Break aim iteration draw, reached max iteration count aim.aim_pos({:?}) aim_vel({:?})", aim_target.aim_pos, aim_target.aim_vel);
+                    warn!(message);
+                    break;
+                }
+                if let Some((entity, hit)) = rapier_context.cast_shape(
+                    ray_start,
+                    0.0,
+                    ray_vel,
+                    &shape,
+                    CAST_RAY_MAX_TOI,
+                    kinematic_filter,
+                ) {
+                    if let Ok(_) = wall_query.get(entity) {
+                        let center = (hit.witness1 - hit.witness2).add(Vec2::new(
                             0.0,
-                            ray_vel,
-                            &shape,
-                            CAST_RAY_MAX_TOI,
-                            kinematic_filter,
-                        ) {
-                            // println!("entity_id({}) hit {:?}", entity.index(), hit);
-                            if let Ok(_) = wall_query.get(entity) {
-                                let center = hit.witness1 - hit.witness2;
-                                commands.spawn(AimBundle::new_line(
-                                    ray_start,
-                                    center,
-                                    &mut meshes,
-                                    &gameplay_materials,
-                                ));
-                                ray_start = center;
-                                ray_vel = Vec2::new(-ray_vel.x, ray_vel.y);
-                            } else if let Ok(_) = balls_query.get(entity) {
-                                let center = hit.witness1 - hit.witness2;
-                                target_transform.translation =
-                                    center.extend(target_transform.translation.z);
-                                commands.spawn(AimBundle::new_line(
-                                    ray_start,
-                                    center,
-                                    &mut meshes,
-                                    &gameplay_materials,
-                                ));
-                                break;
-                            } else if let Ok(_) = lines_query.get(entity) {
-                                let center = hit.witness1 - hit.witness2;
-                                target_transform.translation =
-                                    center.extend(target_transform.translation.z);
-                                commands.spawn(AimBundle::new_line(
-                                    ray_start,
-                                    center,
-                                    &mut meshes,
-                                    &gameplay_materials,
-                                ));
-                                break;
-                            }
+                            CAST_RAY_BOUNCE_Y_ADD * ray_vel.normalize().y,
+                        ));
+                        commands.spawn(AimBundle::new_line(
+                            ray_start,
+                            center,
+                            &mut meshes,
+                            &gameplay_materials,
+                        ));
+                        ray_start = center;
+                        let mut reverse_x_vel = -ray_vel.x;
+                        match reverse_x_vel.signum() > 0.0 {
+                            true => reverse_x_vel += CAST_RAY_VELOCITY_TOLERANCE,
+                            false => reverse_x_vel -= CAST_RAY_VELOCITY_TOLERANCE,
                         }
+                        ray_vel = Vec2::new(reverse_x_vel, ray_vel.y);
+                    } else if let Ok(_) = balls_query.get(entity) {
+                        let center = hit.witness1 - hit.witness2;
+                        target_transform.translation =
+                            center.extend(target_transform.translation.z);
+                        commands.spawn(AimBundle::new_line(
+                            ray_start,
+                            center,
+                            &mut meshes,
+                            &gameplay_materials,
+                        ));
+                        *target_visibility = Visibility::Visible;
+                        break;
+                    } else if let Ok(_) = lines_query.get(entity) {
+                        let center = hit.witness1 - hit.witness2;
+                        target_transform.translation =
+                            center.extend(target_transform.translation.z);
+                        commands.spawn(AimBundle::new_line(
+                            ray_start,
+                            center,
+                            &mut meshes,
+                            &gameplay_materials,
+                        ));
+                        *target_visibility = Visibility::Visible;
+                        break;
                     }
-                } else {
-                    *target_visibility = Visibility::Hidden;
                 }
             }
         }
